requirements.txt
django
djangorestframework
django-cors-headers
# uwsgi
stripe
uuid
requests
